Old project's new to-dos are saved specifically.
If the user creates a new project, with to-dos, and later selects the same project again, the old to-dos are saved and the user is able to create new to-dos which are saved on their specifically array.
Now i will focus on returning their respective content, matching their objects with their specific elements.