Fixed old projects not returning to-do changes.
I was pushing the already existing objects, so they were getting duplicated, so i changed the logic to change the existing details.
In the other case, when i need to create new to-dos objects, we can't change existing details, so in this case, we have to push them to the object list.