Added a way to expand to-dos and change details.
I had to create the details of the to-dos inside of the genToDosDOM function, in order to reutilize the logic previously done. So in this case i couldn't generate new to-dos each time i want to expand a to-do.
What i did is to having these elements hidden by default, using width and height 0, so when i push the button to edit them, they have a normal width and height, so they are visible.
I did it in this way because if i errased the elements directly, i would have to wire up tons of new logic to save the details to display them later, so hidding them allows me to reutilize the old logic that i used previously to save up the to-do title.